requirements.txt
PyGithub>=1.55.0
python-dotenv>=0.19.0
pandas>=1.3.0
requests>=2.25.0

# AI library for PR summarization
openai>=1.0.0